src/procurement_granite/data/requirements.txt
requests
pdfplumber
pymupdf
python-docx
tqdm
pdf2image
pytesseract
tika